dashboard/requirements.txt
psycopg2-binary
fastapi
uvicorn
websockets
datetime
